Refactor game constants and update player time handling for improved gameplay mechanics

+++ inc/player.h
@@ -6,7 +6,7 @@
 /*   By: enchevri <enchevri@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/29 13:27:37 by gabach            #+#    #+#             */
-/*   Updated: 2025/11/30 14:21:50 by enchevri         ###   ########lyon.fr   */
+/*   Updated: 2025/11/30 17:15:46 by enchevri         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ typedef struct s_player
 	char				direction;
 	int					lives;
 	int					score;
-	double				time;
+	int					time;
 	int					invincibility;
 }						t_player;
 
@@ -52,6 +52,7 @@ int						game_loop(t_game *game);
 t_player				init_player(void);
 void					player_action(int c, t_player *player,
 							t_list **projectile);
-void	render_player(t_player *player, t_list **projectiles, t_list **oponents);
+void					render_player(t_player *player, t_list **projectiles,
+							t_list **oponents);
 
 #endif
